+++ packages/engine/Source/Scene/QuadtreePrimitive.js
@@ -283,13 +283,17 @@ QuadtreePrimitive.prototype.updateHeight = function (cartographic, callback) {
   object.removeFunc = function () {
     const addedCallbacks = primitive._addHeightCallbacks;
     const length = addedCallbacks.length;
+    let stillWaiting = false;
     for (let i = 0; i < length; ++i) {
       if (addedCallbacks[i] === object) {
         addedCallbacks.splice(i, 1);
+        stillWaiting = true;
         break;
       }
     }
-    primitive._removeHeightCallbacks.push(object);
+    if (!stillWaiting) {
+      primitive._removeHeightCallbacks.push(object);
+    }
     if (object.callback) {
       object.callback = undefined;
     }